Implement wall collisions
For now the only walls are the implicit canvas boundaries. I will
add in-map walls in another commit. We might decide that we don't
want these boundaries but we'd need to figure out camera logic and
all that so for now I'm doing this.

The logic isn't perfect but works for now.